+++ cpp/GlobalObject.h
@@ -67,6 +67,8 @@ public:
     Q_INVOKABLE QString checkMatrixServerUrl(QString url);
     Q_INVOKABLE AndroidMaterialTheme materialColors();
     Q_INVOKABLE QString mediaCachePath();
+    Q_INVOKABLE bool mobileMode();
+    Q_INVOKABLE QString themeName();
     
 public slots:
     bool handleMatrixUri(const QByteArray &uri);