+++ cpp/Theme.cpp
@@ -8,11 +8,11 @@
 Q_DECLARE_METATYPE(Theme)
 
 QPalette
-Theme::paletteFromTheme(QStringView theme)
+Theme::paletteFromTheme(Theme::ThemeMode theme)
 {
     [[maybe_unused]] static auto meta = qRegisterMetaType<Theme>("Theme");
     static QPalette original;
-    if (theme == u"light") {
+    if (theme == ThemeMode::Light) {
         static QPalette lightActive = [] {
             QPalette lightActive(
               /*windowText*/ QColor(0x33, 0x33, 0x33),
@@ -34,7 +34,7 @@ Theme::paletteFromTheme(QStringView theme)
             return lightActive;
         }();
         return lightActive;
-    } else if (theme == u"dark") {
+    } else if (theme == ThemeMode::Dark) {
         static QPalette darkActive = [] {
             QPalette darkActive(
               /*windowText*/ QColor(0xca, 0xcc, 0xd1),
@@ -61,16 +61,16 @@ Theme::paletteFromTheme(QStringView theme)
     }
 }
 
-Theme::Theme(QStringView theme)
+Theme::Theme(Theme::ThemeMode theme)
 {
     auto p     = paletteFromTheme(theme);
     separator_ = p.mid().color();
-    if (theme == u"light") {
+    if (theme == ThemeMode::Light) {
         sidebarBackground_ = QColor(0x23, 0x36, 0x49);
         alternateButton_   = QColor(0xcc, 0xcc, 0xcc);
         red_               = QColor(0xa8, 0x23, 0x53);
         orange_            = QColor(0xfc, 0xbe, 0x05);
-    } else if (theme == u"dark") {
+    } else if (theme == ThemeMode::Dark) {
         sidebarBackground_ = QColor(0x2d, 0x31, 0x39);
         alternateButton_   = QColor(0x41, 0x4A, 0x59);
         red_               = QColor(0xa8, 0x23, 0x53);