+++ qml/RoomMembers.qml
@@ -16,7 +16,7 @@ import CursorShape 1.0
 import Crypto 1.0
 import MtxEvent 1.0
 
-ApplicationWindow {
+CustomApplicationWindow {
     id: roomMembersRoot
 
     property MemberList members
@@ -26,8 +26,6 @@ ApplicationWindow {
     height: 650
     width: 420
     minimumHeight: 420
-    palette: GlobalObject.colors
-    color: GlobalObject.colors.window
     flags: Qt.Dialog | Qt.WindowCloseButtonHint | Qt.WindowTitleHint
 
     Shortcut {
@@ -266,6 +264,10 @@ ApplicationWindow {
     footer: DialogButtonBox {
         standardButtons: DialogButtonBox.Ok
         onAccepted: roomMembersRoot.close()
+        background: Rectangle {
+            anchors.fill: parent
+            color: GlobalObject.colors.window
+        }
     }
 
 }