+++ qml/device-verification/DeviceVerification.qml
@@ -8,8 +8,9 @@ import QtQuick.Controls 2.3
 import QtQuick.Window 2.13
 import VerificationManager 1.0
 import GlobalObject 1.0
+import "../"
 
-ApplicationWindow {
+CustomApplicationWindow {
     id: dialog
 
     property var flow
@@ -17,18 +18,11 @@ ApplicationWindow {
     onClosing: VerificationManager.removeVerificationFlow(flow)
     title: stack.currentItem ? (stack.currentItem.title_ || "") : ""
     modality: Qt.NonModal
-    palette: GlobalObject.colors
-    color: GlobalObject.colors.window
     //height: stack.currentItem.implicitHeight
     minimumHeight: stack.currentItem.implicitHeight + 2 * GlobalObject.paddingLarge
     height: stack.currentItem.implicitHeight + 2 * GlobalObject.paddingMedium
     minimumWidth: 400
     flags: Qt.Dialog | Qt.WindowCloseButtonHint | Qt.WindowTitleHint
-
-    background: Rectangle {
-        color: GlobalObject.colors.window
-    }
-
 
     StackView {
         id: stack