+++ qml/ui/dialogs/RawMessageDialog.qml
@@ -6,16 +6,15 @@
 import QtQuick 2.15
 import QtQuick.Controls 2.15
 import GlobalObject 1.0
+import "../../"
 
-ApplicationWindow {
+CustomApplicationWindow {
     id: rawMessageRoot
 
     property alias rawMessage: rawMessageView.text
 
     height: 420
     width: 420
-    palette: GlobalObject.colors
-    color: GlobalObject.colors.window
     flags: Qt.Tool | Qt.WindowStaysOnTopHint | Qt.WindowCloseButtonHint | Qt.WindowTitleHint
 
     Shortcut {
@@ -50,6 +49,10 @@ ApplicationWindow {
     footer: DialogButtonBox {
         standardButtons: DialogButtonBox.Ok
         onAccepted: rawMessageRoot.close()
+        background: Rectangle {
+            anchors.fill: parent
+            color: GlobalObject.colors.window
+        }
     }
 
 }